requirements.txt
flask
flask-cors
faiss-cpu
numpy
langchain-google-genai
huggingface_hub


